requirements.txt
alembic==1.10.4
blinker==1.6.2
click==8.1.3
decorator==5.1.1
Deprecated==1.2.13
exceptiongroup==1.1.1
Flask==2.3.2
flask-blueprint==1.3.0
Flask-Gunicorn==0.1.1
Flask-Limiter==3.3.1
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==2.0.2
gunicorn==19.9.0
importlib-resources==5.12.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
limits==3.4.0
Mako==1.2.4
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
ordered-set==4.1.0
packaging==23.1
pluggy==1.0.0
Pygments==2.15.1
pypng==0.20220715.0
pytest==7.4.0
pytest-flask==1.2.0
python-dotenv==1.0.0
qrcode==7.4.2
rich==13.3.5
SQLAlchemy==2.0.17
tomli==2.0.1
typing_extensions==4.5.0
Werkzeug==2.3.3
wrapt==1.15.0
WTForms==3.0.1
